# yugo/openapi.yaml
...
openapi: 3.1.0
info:
  title: Yugo Body API
  version: 0.1.0
  description: |
    **The body** is Yugo's nervous system: a single light FastAPI on the dog's
    LAN that owns the one WebRTC `LocalSTA` link to the Go2 Air, executes all
    control locally and fast, serves the app, and runs the agent loop — but
    **delegates every intelligence step to the mind (cloud)**.

    - **Body (this API)** = robot I/O, local control, reflex/safety loop, state
      aggregation, app contract.
    - **Mind (cloud)** = LLM reasoning (OpenAI), perception (vision/detectors),
      voice & music generation (ElevenLabs). The body carries *text*, not audio.

    Routes are tagged into groups (control, telemetry, agent, sensor, audio,
    system) and each operation declares `x-execution`:
    `local` (runs over WebRTC on the robot) or `delegates-to-mind` (calls cloud).

    ## No auth — LAN tool
    This is a local LAN/Tailscale tool and ships with **no authentication** by
    design. Do **not** expose its port to the public internet. An empty,
    optional security scheme is declared (`security: [{}]`) purely to make the
    "no auth required" contract explicit for code generators.

    ## Three hard constraints (from live testing 2026-05-28/29)
    1. **BalanceStand precondition.** Expressive `SPORT_CMD` moves (e.g.
       `WiggleHips`) are *ignored* unless Yugo is upright. The body gates them
       behind an auto-`BalanceStand`. `Hello` works from most states.
...

    ## Degradation
...
servers:
...
security:
  - {}
...
paths:
...
  /feed:
    get:
      tags: [telemetry]
      operationId: getFeed
      summary: MJPEG camera stream (hub)
      x-execution: local
      description: |
        **IMPLEMENTED** (hub `FeedRouter` + `FeedRelay`). The camera on the hub
        (:8080), superseding the deprecated WebBridge `/video_feed/color_image`
        (:5555). Serves the dog's color camera as a continuous
        `multipart/x-mixed-replace` MJPEG stream (boundary `frame`, parts
        `image/jpeg`). Frames are sourced from the dog's decoded video stream;
        `YUGO_FEED_FAKE=1` substitutes synthetic color bars (no dog).

        **Offline / no frame source:** returns `503` ("no video source"). The
        sibling WebRTC path (`POST /feed/offer`) serves the same frames to a
        viewer's peer connection; the cockpit viewer is at `GET /feed/cockpit`.

        **Safety:** read-only; non-critical telemetry (shares no state with the
        deadman/reflex layer).
      responses:
        '200':
          description: An open MJPEG multipart stream.
          content:
            multipart/x-mixed-replace:
              schema:
                type: string
                format: binary
                description: "Repeating `--frame` parts, each `Content-Type: image/jpeg`."
        '503':
          description: No frame source (offline / no synthetic source).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /feed/offer:
    post:
      tags: [telemetry]
      operationId: postFeedOffer
      summary: WebRTC signaling (WHIP-style) → SDP answer
      x-execution: local
      description: |
        **IMPLEMENTED** (hub `FeedRouter` + `FeedRelay`). Single-shot,
        non-trickle signaling: the viewer POSTs its SDP offer (with host ICE
        candidates bundled); the hub creates a peer connection, attaches the
        live camera track, and returns its SDP answer. LAN-only (no STUN/TURN).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SdpOffer' }
      responses:
        '200':
          description: SDP answer.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SdpAnswer' }
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          description: No frame source (offline / no synthetic source).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '504':
          description: Signaling/ICE did not complete in time; retry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
...
  responses:
...
    ValidationError:
      description: 422 — request body failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            invalid:
              value: { ok: false, error: "invalid mode 'sleepy'" }

  # ---- Schemas -------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      description: Standard error envelope.
      required: [ok, error]
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: string
          description: Human-readable error message.
...
    SdpOffer:
      type: object
      description: WHIP-style WebRTC offer (the `POST /feed/offer` body).
      required: [sdp]
      properties:
        sdp: { type: string, description: The viewer's SDP offer. }
        type: { type: string, enum: [offer], default: offer }
    SdpAnswer:
      type: object
      description: The hub's SDP answer returned from `POST /feed/offer`.
      required: [sdp, type]
      properties:
        sdp: { type: string, description: The hub's SDP answer. }
        type: { type: string, enum: [answer] }
...